Tabulator: generalized auto-tabulation extras -> dict/Dict; moved tabulator.get_nested_* methods to node.util
